@@ -1,6 +1,5 @@
 ﻿using System;
-using System.Data.SqlClient;
-using System.Diagnostics;
+using Microsoft.Data.SqlClient;
 using System.Linq;
 
 namespace Loudenvier.Utils;
