@@ -89,7 +89,7 @@ public class Crc64 : HashAlgorithm
                 if ((entry & 1) == 1)
                     entry = (entry >> 1) ^ polynomial;
                 else
-                    entry = entry >> 1;
+                    entry >>= 1;
             createTable[i] = entry;
         }
         return createTable;
